requirements.txt
python-telegram-bot>=20.6
yt-dlp>=2024.10.22
